pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.13.0",
    "asyncpg>=0.30.0",
    "fastapi[all]>=0.118.2",
    "pydantic-settings>=2.11.0",
    "redis[async]>=6.4.0",
    "rich>=14.1.0",
    "sqlalchemy>=2.0.43",
]

[dependency-groups]
dev = [
    "ruff>=0.14.0",
]
